support static model defs
